+++ gridstatus/tests/test_spp.py
@@ -368,7 +368,7 @@ class TestSPP(BaseTestISO):
         assert df["Interval Start"].max().date() == two_days_ago.date()
         self._check_ver_curtailments(df)
 
-    def test_get_get_ver_curtailments_annual(self):
+    def test_get_ver_curtailments_annual(self):
         year = 2020
         df = self.iso.get_ver_curtailments_annual(year=year)
 
@@ -376,3 +376,50 @@ class TestSPP(BaseTestISO):
         assert df["Interval Start"].max().date() == pd.Timestamp(f"{year}-12-31").date()
 
         self._check_ver_curtailments(df)
+
+    # get_capacity_of_generation_on_outage
+
+    def _check_capacity_of_generation_on_outage(self, df):
+        columns = [
+            "Publish Time",
+            "Interval Start",
+            "Interval End",
+            "Total Outaged MW",
+            "Coal MW",
+            "Diesel Fuel Oil MW",
+            "Hydro MW",
+            "Natural Gas MW",
+            "Nuclear MW",
+            "Solar MW",
+            "Waste Disposal MW",
+            "Wind MW",
+            "Waste Heat MW",
+            "Other MW",
+        ]
+
+        assert df.columns.tolist() == columns
+
+    def test_get_capacity_of_generation_on_outage(self):
+        two_days_ago = pd.Timestamp.now() - pd.Timedelta(days=2)
+        start = two_days_ago - pd.Timedelta(days=2)
+        df = self.iso.get_capacity_of_generation_on_outage(
+            start=start,
+            end=two_days_ago,
+        )
+
+        self._check_capacity_of_generation_on_outage(df)
+
+        # confirm three weeks of data
+        assert df.shape[0] / 168 == 3
+        assert df["Publish Time"].dt.date.nunique() == 3
+
+    def test_get_capacity_of_generation_on_outage_annual(self):
+        year = 2020
+        df = self.iso.get_capacity_of_generation_on_outage_annual(year=year)
+
+        assert df["Interval Start"].min().date() == pd.Timestamp(f"{year}-01-01").date()
+
+        # 2020 was a leap year
+        assert df["Publish Time"].nunique() == 366
+
+        self._check_capacity_of_generation_on_outage(df)