+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## vNext
+
+### ERCOT
+
+- Add suport for RTM and DAM price correction datasets
+
+### EIA
+
+- Add `EIA.get_grid_monitor` dataset with hourly BA and Region emission data
+
 ## v0.23.0 - Sept 12, 2023
 
 ### ERCOT
@@ -18,7 +28,6 @@
 
 ### EIA
 
-- Add `EIA.get_grid_monitor` dataset with hourly BA and Region emission data
 - Add wholesale petroleum and natural gas daily spot prices.
 - Add weekly spot prices and export totals for coal.
 - Add handler for hourly fuel type data