+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Add `Ercot.get_60_day_sced_disclosure`
 - Add `Ercot.get_60_day_dam_disclosure`
 - Add support for specifying `forecast_type` to `Ercot.get_load_forecast`
+- Add ERCOT System Lambda
 
 ### MISO