+++ CHANGELOG.md
@@ -3,9 +3,13 @@
 ## vNext
 
 ### ERCOT
-
 - Add suport for RTM and DAM price correction datasets
 - Add System Lambda
+- Add support for RTM and DAM price correction datasets
+
+## SPP
+
+- Add support for generation capacity on outage
 
 ### EIA